refactor: simplify authentication properties
- too many boolean fields to know if authentication is with password
  or with key pair and then if key pair has passphrase
- enum is better suited
- this way there is no need to have an intermediate sealed class
