+++ src/main/kotlin/org/wavescale/sourcesync/configurations/AuthenticationType.kt
@@ -1,0 +1,5 @@
+package org.wavescale.sourcesync.configurations
+
+enum class AuthenticationType(val prettyName: String) {
+    PASSWORD("Password"), KEY_PAIR("Key pair")
+}
